DevBuilding 0.0.13 Улучшен отрисовки объёма Изменён перевод из времени в мировые координаты для CoordinateGrid

+++ Assets/_Scripts/ChartScripts/Entities/CoordinateGrid.cs
@@ -57,7 +57,7 @@ namespace Chart
 
         public DateTime FromXAxisToDate(float x)
         {
-            return zeroPoint + x * step;
+            return zeroPoint + (int)x * step;
         }
 
         public int FromDateToXAxis(DateTime dateTime)
